Apply refactor to immunization
