+++ packages/shared-src/src/models/fhir/Immunization/getQueryToFindUpstreamIds.js
@@ -1,0 +1,72 @@
+export function fromAdministeredVaccines(models, table, id) {
+  const { AdministeredVaccine, Encounter, Patient, ReferenceData, ScheduledVaccine, User } = models;
+
+  switch (table) {
+    case AdministeredVaccine.tableName:
+      return { where: { id } };
+    case Encounter.tableName:
+      return {
+        include: [
+          {
+            model: Encounter,
+            as: 'encounter',
+            where: { id },
+          },
+        ],
+      };
+    case Patient.tableName:
+      return {
+        include: [
+          {
+            model: Encounter,
+            as: 'encounter',
+            include: [
+              {
+                model: Patient,
+                as: 'patient',
+                where: { id },
+              },
+            ],
+          },
+        ],
+      };
+    case ReferenceData.tableName:
+      return {
+        include: [
+          {
+            model: ScheduledVaccine,
+            as: 'scheduledVaccine',
+            include: [
+              {
+                model: ReferenceData,
+                as: 'vaccine',
+                where: { id },
+              },
+            ],
+          },
+        ],
+      };
+    case ScheduledVaccine.tableName:
+      return {
+        include: [
+          {
+            model: ScheduledVaccine,
+            as: 'scheduledVaccine',
+            where: { id },
+          },
+        ],
+      };
+    case User.tableName:
+      return {
+        include: [
+          {
+            model: User,
+            as: 'recorder',
+            where: { id },
+          },
+        ],
+      };
+    default:
+      return null;
+  }
+}
