+++ packages/shared-src/src/models/fhir/Immunization/searchParameters.js
@@ -1,0 +1,13 @@
+import { FHIR_SEARCH_PARAMETERS, FHIR_SEARCH_TOKEN_TYPES } from 'shared/constants/fhir';
+
+export const searchParameters = {
+  patient: {
+    type: FHIR_SEARCH_PARAMETERS.REFERENCE,
+    path: [['patient']],
+  },
+  'vaccine-code': {
+    type: FHIR_SEARCH_PARAMETERS.TOKEN,
+    path: [['vaccineCode', 'coding', '[]']],
+    tokenType: FHIR_SEARCH_TOKEN_TYPES.CODING,
+  },
+};
